feat: Sentinel preferredSlaves option (#370)
* Support for “preferredSlave” when connecting to a Sentinel group for
local reads in a distributed environment. Value can be an array of
objects with required properties “ip” and “port” and option property of
“prio” indicating the priority, defaults to 1, lower numbers preferred
first. If the value is a function the available slaves are passed into
it, expecting a single slave as the response. If no preferred slave is
found a random one is returned from those available.

Updated tests to return bad values in SENTINEL slaves command but still
select the correct one based on the preferredSlave result.

* Update README with details on the `preferredSlaves` option.