fix: pipeline with transactions causes unhandled warnings
Pipeline#exec() is a little tricky: it has a init process when
called first time (indicated by whether Pipeline#nodeifiedPromise
is true). Transaction#exec() is similar to that, it has a init
process that parse the result from Pipeline#exec before returning
to users. However, #nodeifiedPromise is not checked here.

Closes: #883