Handle stream errors thrown immediately
Currently, a stream (ex `net.Socket`) is returned via a Promise, and error
handlers are attached to the stream in the Promise callback.

However, if the errors are thrown immediately (ex with `process.nextTick()`),
the error handlers will not have the chance to attach to the stream, thus
unhandled error events will be thrown.

This PR makes all connectors resolve a factory instead of the created stream,
so the event handlers can be attached in the same tick.

Closes #1236
Closes #1209
Closes #1289